bumped changelogs for 1.1.28
This release contains a new command that will be installed via
lg_common. lg-code-to-command which will really only be useful on the
headnodes, and a new ros node called command_handler_node which will
listen on /command_handler and call the lg-code-to-command executable
with whatever data that's supplied to it.

Also some default changes for lg_earth.