Normalize MPlayer window positioning
The ManagedMplayer's ManagedWindow was not detecting its application window.  By passing the instance name as -name to mplayer, the application will set its classname and awesome can filter it.

The observed problem was that with the recent addition of -loop 0, the
video kept moving horizontally across the screen.  This is presumably an
artifact of setting +x+y in -geometry, so that redundant setting was removed.

Because awesome was not setting the window dimensions as expected, width
and height are still set in -geometry.  This is presumably MPlayer using
size hints or something to force expected dimensions for direct
rendering.