set kiosk=True in test file adhocbrowser creation
The default should be true, but because this is a message type, we need
to manually set it. Inside the adhoc browser we always set kiosk to true
or whatever the user passed to the director message. We never rely on
the default when creating a new instance of AdhocBrowser()